+++ optimizely/optimizely_config.py
@@ -1,4 +1,4 @@
-# Copyright 2020, Optimizely
+# Copyright 2020-2021, Optimizely
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
